made the visualization widget only visualization
