+++ src/app/widgets/racer-line-view/start-line-geometry.util.spec.ts
