+++ src/app/widgets/widget-racer-line/widget-racer-line.component.scss
@@ -10,22 +10,25 @@
   width: 100%;
   height: 100%;
   min-height: 1px;
+  // The canvas, the drawing that overlays it, and the edit overlay are all absolute:
+  // anchor them to the widget rather than to whatever ancestor happens to be positioned.
+  position: relative;
 }
 
 .line-canvas {
   position: absolute;
   width: 100% !important;
-  height: calc(100% - var(--racer-button-row, 68px)) !important;
+  height: calc(100% - var(--racer-button-row, 24%)) !important;
   display: block;
   top: 0;
   left: 0;
 }
 
-// The button row takes a fixed height so its touch targets never shrink with the
+// The button row can take a fixed height so its touch targets never shrink with the
 // widget; the drawing takes whatever is left.
 .filler {
   width: 100%;
-  height: calc(100% - var(--racer-button-row, 68px));
+  height: calc(100% - var(--racer-button-row, 24%));
 }
 
 .buttons-container {
@@ -34,7 +37,7 @@
   justify-content: space-between;
   column-gap: 2%;
   position: relative;
-  height: var(--racer-button-row, 68px);
+  height: var(--racer-button-row, 24%);
   width: 100%;
   padding: 0px 10px 14px 10px;
   z-index: 20;
@@ -120,6 +123,48 @@
   flex: 1 1 auto;
   @include mat.button-overrides((
     filled-container-color: var(--kip-starboard-color),
+    filled-label-text-color: var(--kip-contrast-color)
+  ));
+}
+
+// Kept in the DOM while the drawing has the display, so the canvas keeps its size and
+// its registration with the canvas service across a mode change.
+.line-canvas.hidden {
+  visibility: hidden;
+}
+
+// The drawing sits in the same box the canvas occupies, above the button row.
+.line-view {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: calc(100% - var(--racer-button-row, 24%));
+}
+
+// Mode 5 packs four buttons beside the mode button, so only the middle one is fixed.
+.vmg-select-button {
+  flex: 1 1 auto;
+  @include mat.button-overrides((
+    filled-container-color: var(--kip-grey-color),
+    filled-label-text-color: var(--kip-contrast-color)
+  ));
+}
+
+.vmg-reset-button,
+.vmg-clear-button {
+  flex: none;
+  width: 16%;
+  @include mat.button-overrides((
+    filled-container-color: var(--kip-grey-dim-color),
+    filled-label-text-color: var(--kip-contrast-color)
+  ));
+}
+
+// Throwing the samples away is not something to hit by accident.
+.vmg-clear-button {
+  @include mat.button-overrides((
+    filled-container-color: var(--kip-orange-dim-color),
     filled-label-text-color: var(--kip-contrast-color)
   ));
 }
