+++ src/app/widgets/widget-racer-start-line/widget-racer-start-line.component.scss
@@ -1,5 +1,3 @@
-@use '@angular/material' as mat;
-
 :host {
   overflow: clip;
   display: block;
@@ -7,249 +5,9 @@
   width: 100%;
 }
 
-.line-wrapper {
-  width: 100%;
-  height: 100%;
-  min-height: 1px;
-  // Halo behind the labels so the dashed line does not show through them.
-  --racer-halo: var(--kip-widget-card-background-color, var(--kip-background));
-}
-
-// The button row takes a fixed height so its touch targets never shrink with the
-// widget; the drawing takes whatever is left.
-.filler {
-  width: 100%;
-  height: calc(100% - var(--racer-button-row, 68px));
-}
-
-.line-svg {
-  width: 100%;
-  height: 100%;
+// The drawing takes the whole widget: there is no button row to leave space for.
+racer-line-view {
   display: block;
-  font-family: Roboto, sans-serif;
-}
-
-.viz-title {
-  fill: var(--racer-dim);
-  font-size: 16px;
-}
-
-.start-line {
-  stroke: var(--racer-dim);
-  stroke-width: 4;
-  stroke-dasharray: 10 8;
-}
-
-.line-label {
-  fill: var(--racer-color);
-  stroke: var(--racer-halo);
-  stroke-width: 8;
-  paint-order: stroke;
-}
-
-.line-label {
-  // Sized by the scene, which shrinks it to fit a narrow widget.
-  text-anchor: middle;
-}
-
-
-.port-mark {
-  fill: var(--kip-port-color);
-  stroke: var(--racer-halo);
-  stroke-width: 1;
-}
-
-.stb-mark {
-  fill: var(--kip-starboard-color);
-  stroke: var(--racer-halo);
-  stroke-width: 1;
-}
-
-.projection {
-  stroke: var(--kip-blue-color);
-  stroke-linecap: round;
-
-  // The course actually sailed behind the best VMG. Its own colour again, but purple
-  // now that pink belongs to the approach dimensions - this optional line must not be
-  // mistaken for them.
-  &.best {
-    stroke: var(--kip-purple-color);
-  }
-}
-
-// The start zone boundary: the line's extensions and the 45 degree wedges. Kept well
-// back so the start line itself stays the dominant object.
-.zone-guide {
-  stroke: var(--racer-dim);
-  stroke-width: 1.2;
-  stroke-opacity: 0.6;
-  stroke-dasharray: 4 5;
-}
-
-// The approach legs, drawn the way a drawing dimensions a distance - a thin rule with a
-// tick at each end and its value set into a break in it - so they read as a measurement
-// rather than as a course to steer.
-.dimension {
-  line {
-    stroke: var(--kip-pink-color);
-    stroke-width: 1.5;
-    stroke-opacity: 0.9;
-  }
-
-  text {
-    fill: var(--kip-pink-color);
-    font-size: 15px;
-    font-weight: bold;
-    stroke: var(--racer-halo);
-    stroke-width: 4;
-    paint-order: stroke;
-  }
-}
-
-.dimension-ticks {
-  stroke: var(--kip-pink-color);
-  stroke-width: 1.5;
-  fill: none;
-}
-
-// Where the boat reaches at the gun along those legs.
-.gun-mark circle {
-  fill: var(--racer-halo);
-  stroke: var(--kip-pink-color);
-  stroke-width: 2.5;
-}
-
-.boat {
-  // Two nested hull outlines: even-odd leaves the band between them, whose inner edge is
-  // the vessel at true scale.
-  fill: var(--kip-blue-color);
-  fill-rule: evenodd;
-
-  &.ocs {
-    fill: var(--kip-orange-color);
-  }
-}
-
-.vmg-selection {
-  fill: none;
-  stroke: var(--kip-blue-color);
-  stroke-width: 2;
-}
-
-.vmg-value {
-  // Sized by the scene, which grows it to fill the room the cross has.
-  fill: var(--racer-color);
-  font-weight: bold;
-  text-anchor: middle;
-  stroke: var(--racer-halo);
-  stroke-width: 5;
-  paint-order: stroke;
-
-  // A manually adjusted VMG stands in for the collected samples, so it is called out.
-  &.overridden {
-    fill: var(--kip-orange-color);
-  }
-}
-
-.no-line {
-  fill: var(--racer-dim);
-  font-size: 16px;
-  text-anchor: middle;
-}
-
-.buttons-container {
-  display: flex;
-  flex-wrap: nowrap;
-  justify-content: space-between;
-  column-gap: 2%;
-  position: relative;
-  height: var(--racer-button-row, 68px);
-  width: 100%;
-  padding: 0px 10px 14px 10px;
-  z-index: 20;
-  pointer-events: auto;
-  align-items: center;
-}
-
-.button {
-  min-width: 0px;
-  width: 100%;
-  border-radius: 12px;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  @include mat.button-overrides((
-    filled-container-height: 100%,
-    filled-horizontal-padding: 0px,
-    filled-icon-spacing: 0px,
-  ));
-}
-
-.select-button {
-  flex: none;
-  width: 18%;
-  @include mat.button-overrides((
-    filled-container-color: var(--kip-blue-color),
-    filled-label-text-color: var(--kip-contrast-color)
-  ));
-}
-
-.reset-button {
-  flex: none;
-  width: 12%;
-  @include mat.button-overrides((
-    filled-container-color: var(--kip-grey-dim-color),
-    filled-label-text-color: var(--kip-contrast-color)
-  ));
-}
-
-.vmg-button {
-  flex: 1 1 auto;
-  @include mat.button-overrides((
-    filled-container-color: var(--kip-grey-color),
-    filled-label-text-color: var(--kip-contrast-color)
-  ));
-}
-
-.adjust-button {
-  flex: 1 1 auto;
-  @include mat.button-overrides((
-    filled-container-color: var(--kip-blue-dim-color),
-    filled-label-text-color: var(--kip-contrast-color)
-  ));
-}
-
-.buttons-svg-icon {
-  width: 100%;
   height: 100%;
-  margin: 0px;
-  padding: 0px;
-}
-
-.svg-content {
   width: 100%;
-  height: 100%;
-  padding: 0em;
-  margin: 0em;
-
-  text {
-    font-family: Roboto;
-    font-size: 15px;
-    font-weight: bold;
-    text-align: center;
-    text-anchor: middle;
-    font-style: normal;
-    font-variant: normal;
-    stroke: none;
-  }
-}
-
-.widgetOverlay {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  right: 0;
-  bottom: 0;
-  color: transparent;
-  z-index: 21;
 }
